fix: refactor permissions and buffers
